CUDA: Fix various issues in test suite

+++ numba/cuda/tests/cudadrv/test_linker.py
@@ -78,8 +78,8 @@ class TestLinker(CUDATestCase):
             i = cuda.grid(1)
             x[i] += bar(y[i])
 
-        A = np.array([123])
-        B = np.array([321])
+        A = np.array([123], dtype=np.int32)
+        B = np.array([321], dtype=np.int32)
 
         foo[1, 1](A, B)
